requirements.txt
# Core dependencies
numpy==1.21.0
scipy==1.7.0
Pillow==8.2.0

# Image processing
opencv-python==4.5.2.54
scikit-image==0.18.1

# Machine Learning and Object Detection
tensorflow==2.5.0
torch==1.9.0
torchvision==0.10.0

# GUI
PyQt5==5.15.4

# Cloud Services
boto3==1.17.95  # for AWS
google-cloud-storage==1.38.0
azure-storage-blob==12.8.1

# Social Media Integration
tweepy==3.10.0
facebook-sdk==3.1.0

# Collaboration
python-socketio==5.3.0

# Image Enhancement
rawpy==0.16.0  # for RAW image processing

# Misc Utilities
requests==2.25.1
tqdm==4.61.1  # for progress bars

# Testing
pytest==6.2.4

# Linting and Code Quality
pylint==2.8.3
black==21.6b0

# Documentation
Sphinx==4.0.2

# Logging
loguru==0.5.3

# Performance Optimization
numba==0.53.1

# Version Control (if needed for programmatic access)
GitPython==3.1.18
